# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: CareerRadar API — personal placement OS for CS students
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /saved-searches:
    get:
      operationId: listSavedSearches
      tags: [saved-searches]
      summary: List user's saved searches
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Saved searches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedSearch"
        "401":
          description: Unauthorized
    post:
      operationId: createSavedSearch
      tags: [saved-searches]
      summary: Save a search
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedSearchInput"
      responses:
        "201":
          description: Saved search created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedSearch"
        "401":
          description: Unauthorized
  /saved-searches/{id}:
    delete:
      operationId: deleteSavedSearch
      tags: [saved-searches]
      summary: Delete a saved search
      security:
        - clerkAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    SavedSearch:
      type: object
      required: [id, clerkId, name, filters, createdAt]
      properties:
        id:
          type: string
        clerkId:
          type: string
        name:
          type: string
        filters:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    SavedSearchInput:
      type: object
      required: [name, filters]
      properties:
        name:
          type: string
        filters:
          type: object
          additionalProperties: true

    # ─── Notification ───────────────────────────────────────────────────────────
...